break down Grid into smaller components
