+++ client/components/admin_dashboard/Grid.jsx
@@ -7,42 +7,9 @@ import {
   GridToolbarContainer,
 } from "@mui/x-data-grid";
 import { Paper, Box, Tooltip } from "@mui/material";
-import Button from "@mui/material/Button";
 import SearchBar from "./SearchBar";
-import StatusChange from "./StatusChange";
+import { getColumns } from "./grid-components/columns";
 import { updateStatusOnServer } from "@/actions/form";
-
-const formatName = (name) => {
-  const [firstName, ...rest] = name.split(" ");
-  const lastName = rest.join(" ");
-  if (!lastName) {
-    return firstName;
-  }
-  return `${lastName}, ${firstName}`;
-};
-
-const formatTime = (hour) => {
-  return hour <= 12 ? `${hour}a` : `${hour - 12}p`;
-};
-
-const formatTimeRange = (timeRange) => {
-  return `${formatTime(timeRange?.preferredEarlyTime)} - ${formatTime(
-    timeRange?.preferredLateTime
-  )}`;
-};
-
-const formatDateCreated = (date) => {
-  return `${date.toLocaleDateString()} ${date.toLocaleTimeString([], {
-    hour: "2-digit",
-    minute: "2-digit",
-  })}`;
-};
-
-const formatPhone = (phone) => {
-  const cleaned = ("" + phone).replace(/\D/g, "");
-  const match = cleaned.match(/^(\d{3})(\d{3})(\d{4})$/);
-  return match ? `(${match[1]}) ${match[2]}-${match[3]}` : phone;
-};
 
 const paginationModel = { page: 0, pageSize: 15 };
 
@@ -55,95 +22,15 @@ const Toolbar = () => (
 export default function Grid({ rows, refreshData }) {
   const [searchText, setSearchText] = useState("");
 
-  const columns = [
-    {
-      field: "visitOrder",
-      headerName: "Visit Order",
-      width: 190,
-      valueGetter: (_, row) => row.schedule.order,
-    },
-    {
-      field: "markVisited",
-      headerName: "Visited?",
-      width: 190,
-      renderCell: (params) => {
-        const { id, status } = params.row;
-
-        return (
-          <Button
-            variant={
-              status === "Visited" || status === "Completed"
-                ? "contained"
-                : "outlined"
-            }
-            color="primary"
-            onClick={() =>
-              toggleVisited(
-                id,
-                status === "Visited" || status === "Completed"
-                  ? "Scheduled"
-                  : "Visited"
-              )
-            }
-          >
-            {status === "Visited" || status === "Completed"
-              ? "Visited"
-              : "Not Visited"}
-          </Button>
-        );
-      },
-    },
-    {
-      field: "status",
-      headerName: "Status",
-      width: 190,
-      renderCell: (params) => {
-        const { id, status } = params.row;
-        return (
-          <StatusChange
-            id={id}
-            refreshData={refreshData}
-            currentStatus={status}
-          />
-        );
-      },
-    },
-    {
-      valueFormatter: formatName,
-      field: "name",
-      headerName: "Name",
-      width: 190,
-    },
-    {
-      valueFormatter: formatDateCreated,
-      field: "dateCreated",
-      headerName: "Requested on",
-      width: 190,
-    },
-    {
-      valueFormatter: formatTimeRange,
-      field: "timeRange",
-      headerName: "Timeslot",
-      width: 190,
-    },
-    {
-      valueFormatter: formatPhone,
-      field: "phone",
-      headerName: "Phone",
-      width: 190,
-    },
-    { field: "email", headerName: "Email", width: 190 },
-    {
-      field: "address",
-      headerName: "Address",
-      width: 190,
-    },
-  ];
-
   const toggleVisited = async (id, status) => {
     await updateStatusOnServer(id, status);
     await refreshData();
   };
+
+  const columns = useMemo(
+    () => getColumns(refreshData, toggleVisited),
+    [refreshData]
+  );
 
   const filteredRows = useMemo(() => {
     if (!searchText) return rows;
