+++ client/components/admin_dashboard/grid-components/formatters.jsx
@@ -1,0 +1,31 @@
+export const formatName = (name) => {
+  const [firstName, ...rest] = name.split(" ");
+  const lastName = rest.join(" ");
+  if (!lastName) {
+    return firstName;
+  }
+  return `${lastName}, ${firstName}`;
+};
+
+export const formatTime = (hour) => {
+  return hour <= 12 ? `${hour}a` : `${hour - 12}p`;
+};
+
+export const formatTimeRange = (timeRange) => {
+  return `${formatTime(timeRange?.preferredEarlyTime)} - ${formatTime(
+    timeRange?.preferredLateTime
+  )}`;
+};
+
+export const formatDateCreated = (date) => {
+  return `${date.toLocaleDateString()} ${date.toLocaleTimeString([], {
+    hour: "2-digit",
+    minute: "2-digit",
+  })}`;
+};
+
+export const formatPhone = (phone) => {
+  const cleaned = ("" + phone).replace(/\D/g, "");
+  const match = cleaned.match(/^(\d{3})(\d{3})(\d{4})$/);
+  return match ? `(${match[1]}) ${match[2]}-${match[3]}` : phone;
+};
